fix: Height 생성 시 Players 파라미터 추가
- Ladder 클래스 참여자의 수 ~ 참여자의 수 * 2 까지 검증 로직 위임
